Legg mock perioder i egen fil som kan slettes senere
